Converts HomeScreen to react effects for redux
