--- TTKDTrainingApp/src/components/HomeScreen/index.js
@@ -1,3 +1,0 @@
-import {HomeScreenConnector} from './HomeScreenConnector';
-
-export default HomeScreenConnector;
